Soil Update Type Updates Removed Windows Header in PlatformCore Refactor D3D12 Device Handling and Sync Warning Fixes

+++ GraphicsEngine/Source/Core/Core/Utils/NVAPIManager.cpp
@@ -65,7 +65,7 @@ void NVAPIManager::RenderGPUStats(int statx, int starty)
 	{
 		for (int i = 0; i < GpuData[x].size(); i++)
 		{
-			TextRenderer::instance->RenderFromAtlas(GpuData[x][i], (float)statx + x * Xsize, starty - Ysize * i, 0.35f);
+			TextRenderer::instance->RenderFromAtlas(GpuData[x][i], (float)statx + x * Xsize, (float)starty - Ysize * i, 0.35f);
 		}
 	}
